+++ PF_Api/package.json
@@ -15,13 +15,12 @@
     "npm": ">=6.14.6"
   },
   "dependencies": {
-    "@mercadopago/sdk-react": "^0.0.3",
     "axios": "^1.3.2",
     "body-parser": "^1.19.0",
     "cookie-parser": "^1.4.5",
     "cors": "^2.8.5",
     "express": "^4.18.2",
-    "mercadopago": "^0.0.6",
+    "mercadopago": "^1.5.15",
     "morgan": "^1.10.0",
     "multer": "^1.4.5-lts.1",
     "node-fetch": "^3.3.0",
