+++ PF_Api/src/handlers/paymentHandlers.js
@@ -1,33 +1,81 @@
+const { CLIENT_ID, CLIENT_SECRET, ACCESS_TOKEN, REDIRECT_URI} = process.env;
+const { User } = require("../db.js");
 const mercadopago = require("mercadopago");
-const { createPreference, sendInformationMP } = require("../controllers/paymentControllers");
+const axios = require('axios');
 
-const postPaymentHandler = (req,res) => {
-  mercadopago.configure({
-    access_token: process.env.ACCESS_TOKEN,
-  })
+const postPaymentHandler = async (req, res) => {
   const { items, seller_id } = req.body;
-  try{
-    res.status(200).json(createPreference(items, seller_id));
+  const seller = await User.findByPk(seller_id);
+  mercadopago.configure({
+    access_token: seller.MPAccessToken,
+  });
+  let preference = {
+    back_urls: {
+      success: "https://my-seam.vercel.app/",
+      failure: "http://localhost:3000/checkout/failure",
+      pending: "http://localhost:3000/checkout/pending",
+    },
+    auto_return: "approved",
+    items: [],
+  };
+  items.forEach((item) => {
+    preference.items.push(item);
+  });
+  try {
+    // Enviar solicitud de preferencia
+    mercadopago.preferences
+      .create(preference)
+      .then(function (response) {
+        res.json({
+          global: response.body.id,
+        });
+      })
+      .catch(function (error) {
+        console.log(error);
+      });
   } catch (error) {
     res.status(400).json({ error: error.message });
   }
-  /* mercadopago.preferences.create(preference)
-    .then(function (response) {
-      res.json({
-        global: response.body.id,
-      })
-    })
-    .catch(function (error) {
-      console.log(error);
-    }); */
+}; 
 
-} 
-
-const getAuthCode = (req,res) => {
-  const { code, status } = req.query;
+const getAuthCode = async (req,res) => {
+  const { code, state } = req.query;
   try {
-    sendInformationMP(code, status)
-    res.status(200).json("Success");
+    await axios
+      .post(
+        "https://api.mercadopago.com/oauth/token",
+        {
+          client_secret: CLIENT_SECRET,
+          client_id: CLIENT_ID,
+          grant_type: "authorization_code",
+          code: code,
+          redirect_uri: REDIRECT_URI 
+        },
+        {
+          headers: {
+            Authorization: `Bearer ${ACCESS_TOKEN}`,
+            "Content-Type": "application/json",
+          },
+        }
+      )
+      .then((response) => {
+        User.update({ 
+          MPAccessToken : response.data.access_token,
+          MPUserId : response.data.user_id,
+          MPRefreshToken : response.data.refresh_token,
+          MPExpiresIn : response.data.expires_in
+        },{
+          where: {
+            id: state
+          }
+        })
+      })
+      .catch((error) => {
+        console.error(error);
+      })
+      .finally(() => {
+        res.redirect('https://my-seam.vercel.app/');
+      })
   } catch (error) {
     res.status(400).json({ error: error.message });
   }
